Change ioutil to C, because all libraries should be dependent on C, not C++
