+++ src/kernel/file/elf/architecture/x32/elf_header_32.cpp
@@ -9,8 +9,6 @@
 
 #include <ioutil.h>
 
-using namespace sys;
-
 namespace elf
 {
 	void* elf_header::getHeader()
