+++ src/kernel/file/elf/architecture/x64/elf64.cpp
@@ -10,8 +10,6 @@
 
 #include <ioutil.h>
 
-using namespace sys;
-
 namespace elf
 {
 	static elf elf::readElf(void* elf)
