+++ src/kernel/file/elf/architecture/x64/elf_header_64.cpp
@@ -9,8 +9,6 @@
 
 #include <ioutil.h>
 
-using namespace sys;
-
 namespace elf
 {
 	void* elf_header::getHeader()
